DataFrame Serializer fix (#2281)
