+++ reflex/components/graphing/plotly.py
@@ -1,10 +1,8 @@
 """Component for displaying a plotly graph."""
 
-import json
 from typing import Any, Dict, List
 
 from reflex.components.component import NoSSRComponent
-from reflex.utils.serializers import serializer
 from reflex.vars import Var
 
 try:
@@ -42,23 +40,3 @@ class Plotly(PlotlyLib):
 
     # If true, the graph will resize when the window is resized.
     use_resize_handler: Var[bool]
-
-
-try:
-    from plotly.graph_objects import Figure
-    from plotly.io import to_json
-
-    @serializer
-    def serialize_figure(figure: Figure) -> list:
-        """Serialize a plotly figure.
-
-        Args:
-            figure: The figure to serialize.
-
-        Returns:
-            The serialized figure.
-        """
-        return json.loads(str(to_json(figure)))["data"]
-
-except ImportError:
-    pass
