+++ tests/components/datadisplay/test_datatable.py
@@ -4,10 +4,9 @@ import pytest
 import reflex as rx
 from reflex.components.datadisplay.datatable import (
     DataTable,
-    serialize_dataframe,  # type: ignore
 )
 from reflex.utils import types
-from reflex.utils.serializers import serialize
+from reflex.utils.serializers import serialize, serialize_dataframe
 
 
 @pytest.mark.parametrize(
