+++ tests/components/graphing/test_plotly.py
@@ -2,8 +2,7 @@ import numpy as np
 import plotly.graph_objects as go
 import pytest
 
-from reflex.components.graphing.plotly import serialize_figure  # type: ignore
-from reflex.utils.serializers import serialize
+from reflex.utils.serializers import serialize, serialize_figure
 
 
 @pytest.fixture
